requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
astroid==2.0.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
beautifulsoup4==4.7.1
blinker==1.4
bs4==0.0.1
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudant==2.10.0
colorama==0.4.0
cssselect==1.0.3
Cython==0.29.7
dnspython==1.16.0
docopt==0.6.2
docutils==0.14
dominate==2.3.5
eventlet==0.24.1
fake-useragent==0.1.11
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Passlib==0.1
Flask-Principal==0.4.0
Flask-SocketIO==4.1.0
Flask-WTF==0.14.2
greenlet==0.4.15
httpserver==1.1.0
idna==2.7
idna-ssl==1.1.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.3.2
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
multidict==4.5.2
parse==1.11.1
passlib==1.6.1
pycparser==2.19
pyee==5.0.0
Pygments==2.3.0
pylint==2.1.1
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-engineio==3.5.1
python-path==0.1.3
python-socketio==3.1.2
pythonanywhere==0.7.4
recordtype==1.3
requests==2.20.1
requests-html==0.10.0
simple-http-server==0.1.7
six==1.11.0
slackclient==1.3.0
soupsieve==1.8
tqdm==4.31.1
typing-extensions==3.7.2
urllib3==1.24.1
urlparse2==1.1.1
visitor==0.1.3
w3lib==1.20.0
watson-developer-cloud==2.4.1
websocket-client==0.47.0
websockets==7.0
Werkzeug==0.14.1
wrapt==1.10.11
WSGIUtils==0.7.1
WTForms==2.2.1
WTForms-JSON==0.3.3
yarl==1.3.0
